requirements.txt
# Core Dependencies
selenium==4.15.2
webdriver-manager==4.0.1

# Image Processing & OCR
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# PDF Generation
pdfkit==1.0.0
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.3

# Web Interface
streamlit==1.28.2

# Utilities
lxml==4.9.3

# Note: You also need to install system dependencies:
# 1. Tesseract OCR: https://github.com/tesseract-ocr/tesseract
# 2. wkhtmltopdf: https://wkhtmltopdf.org/downloads.html